# Conflicts:
#	requirements.txt
#	src/main.py

+++ src/main.py
@@ -10,13 +10,15 @@ import calendar
 from slugify import slugify
 import re
 
+import conf
+conf.LOG
+
 LOG = logging.getLogger(__name__)
 LOG.addHandler(logging.FileHandler('scrape.log'))
 LOG.level = logging.INFO
 
 placeholder_version = 1
 
-placeholder_image_alt = ""
 placeholder_box_title_if_missing = "Placeholder box title because we must have one"
 #
 # utils
@@ -44,41 +46,41 @@ def nonxml(msg):
     "we're scraping a value that doesn't appear in the XML"
     return note("nonxml: %s" % msg, logging.WARN)
 
+#
+#
+#
+
 def display_channel_to_article_type(display_channel_list):
-    types = {}
-    types["Correction"] = "correction"
-    types["Editorial"] = "editorial"
-    types["Feature Article"] = "feature"
-    types["Feature article"] = "feature"
-    types["Insight"] = "insight"
-    types["Registered Report"] = "registered-report"
-    types["Research Advance"] = "research-advance"
-    types["Research Article"] = "research-article"
-    types["Research article"] = "research-article"
-    types["Short report"] = "short-report"
-    types["Tools and Resources"] = "tools-resources"
-    # Note: have not seen the below ones yet, guessing
-    types["Research exchange"] = "research-exchange"
-    types["Retraction"] = "retraction"
-    types["Replication study"] = "replication-study"
-    if display_channel_list:
-        #try:
-        display_channel = display_channel_list[0]
-        #except KeyError:
-        #    display_channel = None
-        if display_channel:
-            for key, value in types.iteritems():
-                if display_channel == key:
-                    return value
+    if not display_channel_list:
+        return    
+    types = {
+        "Correction": "correction",
+        "Editorial": "editorial",
+        "Feature Article": "feature",
+        "Feature article": "feature",
+        "Insight": "insight",
+        "Registered Report": "registered-report",
+        "Research Advance": "research-advance",
+        "Research Article": "research-article",
+        "Research article": "research-article",
+        "Short report": "short-report",
+        "Tools and Resources": "tools-resources",
+        
+        # NOTE: have not seen the below ones yet, guessing
+        "Research exchange": "research-exchange",
+        "Retraction": "retraction",
+        "Replication study": "replication-study",
+    }
+    display_channel = display_channel_list[0]
+    return types.get(display_channel)
 
 def license_url_to_license(license_url):
-    if license_url:
-        if license_url == "http://creativecommons.org/licenses/by/3.0/":
-            return "CC-BY-3.0"
-        if license_url == "http://creativecommons.org/licenses/by/4.0/":
-            return "CC-BY-4.0"
-        if license_url == "http://creativecommons.org/publicdomain/zero/1.0/":
-            return "CC0-1.0"
+    idx = {
+        "http://creativecommons.org/licenses/by/3.0/": "CC-BY-3.0",
+        "http://creativecommons.org/licenses/by/4.0/": "CC-BY-4.0",
+        "http://creativecommons.org/publicdomain/zero/1.0/": "CC0-1.0"
+    }
+    return idx.get(license_url)
 
 def related_article_to_related_articles(related_article_list):
     related_articles = []
@@ -95,23 +97,11 @@ def related_article_to_related_articles(related_article_list):
     return related_articles
 
 def is_poa_to_status(is_poa):
-    if is_poa is True:
-        return "poa"
-    elif is_poa is False:
-        return "vor"
-    return None
+    return "poa" if is_poa else "vor"
 
 def self_uri_to_pdf(self_uri_list):
     if self_uri_list:
         return self_uri_list[0]["xlink_href"]
-
-def body_rewrite(body):
-    body = image_uri_rewrite(body)
-    body = mathml_rewrite(body)
-    body = fix_section_id_if_missing(body)
-    body = fix_paragraph_with_content(body)
-    body = fix_box_title_if_missing(body)
-    return body
 
 def generate_section_id():
     """section id attribute generator"""
@@ -228,6 +218,14 @@ def fix_section_id_if_missing(body_json):
 
     return body_json
 
+def body_rewrite(body):
+    body = image_uri_rewrite(body)
+    body = mathml_rewrite(body)
+    body = fix_section_id_if_missing(body)
+    body = fix_paragraph_with_content(body)
+    body = fix_box_title_if_missing(body)
+    return body
+
 #
 #
 #
@@ -247,14 +245,16 @@ def jats(funcname, *args, **kwargs):
 def category_codes(cat_list):
     return [slugify(cat, stopwords=['and']) for cat in cat_list]
 
+THIS_YEAR = time.gmtime()[0]
 def to_volume(volume):
     if not volume:
         # No volume on unpublished PoA articles, calculate based on current year
-        volume = time.gmtime()[0] - 2011
+        volume = THIS_YEAR - 2011
     return int(volume)
 
 def clean(article_data):
     # Remove null or blank elements
+
     article_json = article_data # we're dealing with json just yet ...
     remove_if_none = ["pdf", "relatedArticles", "digest", "abstract"]
     for remove_index in remove_if_none:
@@ -274,8 +274,9 @@ def clean(article_data):
 
     remove_from_copyright_if_none = ["holder"]
     for remove_index in remove_from_copyright_if_none:
-        if article_json["article"]["copyright"][remove_index] is None:
-            del article_json["article"]["copyright"][remove_index]
+        if article_json["article"].get("copyright", {}).has_key(remove_index):
+            if article_json["article"]["copyright"][remove_index] is None:
+                del article_json["article"]["copyright"][remove_index]
 
     # If abstract has no DOI, turn it into an impact statement
     if "abstract" in article_json["article"] and "impactStatement" not in article_json["article"]:
@@ -300,72 +301,62 @@ def authors_rewrite(authors):
 # 
 #
 
-POA = OrderedDict([
-    ('journal', OrderedDict([
-        ('id', [jats('journal_id')]),
-        ('title', [jats('journal_title')]),
-        ('issn', [jats('journal_issn', 'electronic')]),
+JOURNAL = OrderedDict([
+    ('id', [jats('journal_id')]),
+    ('title', [jats('journal_title')]),
+    ('issn', [jats('journal_issn', 'electronic')]),
+])
+
+SNIPPET = OrderedDict([
+    ('status', [jats('is_poa'), is_poa_to_status]), # shared by both POA and VOR snippets but not obvious in schema
+    ('id', [jats('publisher_id')]),
+    ('version', [placeholder_version, todo('version')]),
+    ('type', [jats('display_channel'), display_channel_to_article_type]),
+    ('doi', [jats('doi')]),
+    ('authorLine', [jats('author_line')]),
+    ('title', [jats('title')]),
+    ('published', [jats('pub_date'), to_isoformat]),
+    ('volume', [jats('volume'), to_volume]),
+    ('elocationId', [jats('elocation_id')]),
+    ('pdf', [jats('self_uri'), self_uri_to_pdf]),
+    ('subjects', [jats('category'), category_codes]),
+    ('research-organisms', [jats('research_organism')]),
+    ('abstract', [jats('abstract_json')]),
+])
+# https://github.com/elifesciences/api-raml/blob/develop/dist/model/article-poa.v1.json#L689
+POA_SNIPPET = copy.deepcopy(SNIPPET)
+
+POA = copy.deepcopy(POA_SNIPPET)
+POA.update(OrderedDict([
+    ('copyright', OrderedDict([
+        ('license', [jats('license_url'), license_url_to_license]),
+        ('holder', [jats('copyright_holder')]),
+        ('statement', [jats('license')]),
     ])),
-    ('article', OrderedDict([
-        ('status', [jats('is_poa'), is_poa_to_status]),
-        ('id', [jats('publisher_id')]),
-        ('version', [placeholder_version, todo('version')]),
-        ('type', [jats('display_channel'), display_channel_to_article_type]),
-        ('doi', [jats('doi')]),
-        ('title', [jats('title')]),
-        ('published', [jats('pub_date'), to_isoformat]),
-        ('volume', [jats('volume'), to_volume]),
-        ('elocationId', [jats('elocation_id')]),
-        ('pdf', [jats('self_uri'), self_uri_to_pdf]),
-        ('subjects', [jats('category'), category_codes]),
-        ('researchOrganisms', [jats('full_research_organism')]),
-        ('abstract', [jats('abstract_json')]),
+    ('authors', [jats('authors_json'), authors_rewrite])
+]))
 
-        # non-snippet values
+VOR_SNIPPET = copy.deepcopy(POA_SNIPPET)
+VOR_SNIPPET.update(OrderedDict([
+    ('impactStatement', [jats('impact_statement')]),
+]))
 
-        ('copyright', OrderedDict([
-            ('license', [jats('license_url'), license_url_to_license]),
-            ('holder', [jats('copyright_holder')]),
-            ('statement', [jats('license')]),
-        ])),
-        ('authorLine', [jats('author_line')]),
-        ('authors', [jats('authors_json'), authors_rewrite])
-        
+VOR = copy.deepcopy(VOR_SNIPPET)
+VOR.update(OrderedDict([
+    ('keywords', [jats('keywords')]),
+    ('relatedArticles', [jats('related_article'), related_article_to_related_articles]),
+    ('digest', [jats('digest_json')]),
+    ('body', [jats('body'), body_rewrite]), # ha! so easy ...
+    ('decisionLetter', [jats('decision_letter'), body_rewrite]),
+    ('authorResponse', [jats('author_response'), body_rewrite]),
+]))
+
+def mkdescription(poa=True):
+    return OrderedDict([
+        ('journal', JOURNAL),
+        ('snippet', POA_SNIPPET if poa else VOR_SNIPPET),
+        ('article', POA if poa else VOR),      
     ])
-)])
-
-
-VOR = copy.deepcopy(POA)
-VOR['article'].update(OrderedDict([
-        ('impactStatement', [jats('impact_statement')]),
-        ('keywords', [jats('full_keywords')]),
-        ('relatedArticles', [jats('related_article'), related_article_to_related_articles]),
-        ('digest', [jats('digest_json')]),
-        ('body', [jats('body'), wrap_body_rewrite]), # ha! so easy ...
-        ('decisionLetter', [jats('decision_letter'), body_rewrite]),
-        ('authorResponse', [jats('author_response'), body_rewrite]),
-]))
-
-# if has attached image ...
-VOR['article'].update(OrderedDict([
-    ('image', OrderedDict([
-        ('alt', [placeholder_image_alt, todo('image alt')]),
-        ('sizes', OrderedDict([
-            ("2:1", OrderedDict([
-                ("900", ["https://...", todo("vor article image sizes 2:1 900")]),
-                ("1800", ["https://...", todo("vor article image sizes 2:1 1800")]),
-            ])),
-            ("16:9", OrderedDict([
-                ("250", ["https://...", todo("vor article image sizes 16:9 250")]),
-                ("500", ["https://...", todo("vor article image sizes 16:9 500")]),
-            ])),
-            ("1:1", OrderedDict([
-                ("70", ["https://...", todo("vor article image sizes 1:1 70")]),
-                ("140", ["https://...", todo("vor article image sizes 1:1 140")]),
-            ])),
-        ])),
-    ]))
-]))
 
 #
 # bootstrap
@@ -373,7 +364,7 @@ VOR['article'].update(OrderedDict([
 
 def render_single(doc):
     soup = to_soup(doc)
-    description = POA if parseJATS.is_poa(soup) else VOR
+    description = mkdescription(parseJATS.is_poa(soup))
     return clean(render(description, [soup])[0])
 
 def main(doc):
@@ -384,4 +375,8 @@ def main(doc):
         raise
 
 if __name__ == '__main__':  # pragma: no cover
+    args = sys.argv[1:]
+    if len(args) == 0:
+        print "path to an article xml file required"
+        exit(1)
     main(sys.argv[1]) # pragma: no cover
